Refactor Query with Module::ClassMethods
