+++ app/classes/query/modules/active_record.rb
@@ -4,53 +4,6 @@
 module Query::Modules::ActiveRecord
   attr_accessor :record
 
-  def self.included(base)
-    base.extend(ClassMethods)
-  end
-
-  # Class methods.
-  module ClassMethods
-    def safe_find(id)
-      find(id)
-    rescue ::ActiveRecord::RecordNotFound
-      nil
-    end
-
-    def find(id)
-      record = QueryRecord.find(id)
-      query = Query.rebuild_from_description(record.description)
-      record.query = query
-      query.record = record
-      QueryRecord.cleanup
-      query
-    end
-
-    def lookup_and_save(*)
-      query = lookup(*)
-      query.record.save!
-      query
-    end
-
-    def lookup(*)
-      query = Query.new(*)
-      record = get_record(query)
-      record.query = query
-      query.record = record
-      QueryRecord.cleanup
-      query
-    end
-
-    def get_record(query)
-      desc = query.serialize
-      QueryRecord.find_by(description: desc) ||
-        QueryRecord.new(
-          description: desc,
-          updated_at: Time.zone.now,
-          access_count: 0
-        )
-    end
-  end
-
   def record
     # This errors out if @record is not set since it
     # cannot find Query.get_record.  If you copy the
